+++ app/ante/evm/eip712.go
@@ -13,7 +13,7 @@
 //
 // You should have received a copy of the GNU Lesser General Public License
 // along with the Ethermint library. If not, see https://github.com/evmos/evmos/blob/main/LICENSE
-package ethante
+package evm
 
 import (
 	"fmt"
