+++ app/ante/evm/fee_checker_test.go
@@ -1,4 +1,4 @@
-package ethante
+package evm
 
 import (
 	"math/big"
